warning added to reset game data button

+++ Hospital Saviour/Assets/Scripts/MenuSelection.cs
@@ -4,6 +4,8 @@ using UnityEngine;
 using UnityEngine.UI;
 //https://www.c-sharpcorner.com/article/unity-change-scene-on-button-click-using-c-sharp-scripts-in-unity/
 using UnityEngine.SceneManagement;
+//https://www.kindacode.com/snippet/unity-how-to-show-a-confirmation-dialog/ accessed 3/9/23
+using UnityEditor;
 
 public class MenuSelection : MonoBehaviour
 {
@@ -135,8 +137,20 @@ public class MenuSelection : MonoBehaviour
     /// </summary>
     public void ResetGame()
     {
-        PlayerPrefs.DeleteAll();
-        SceneManager.LoadScene("MenuScene");
+        //https://www.kindacode.com/snippet/unity-how-to-show-a-confirmation-dialog/ accessed 3/9/23
+        bool decision = EditorUtility.DisplayDialog(
+            "Reset Game", // title
+            "Are you sure want to reset all game data?", // description
+            "Yes", // OK button
+            "No" // Cancel button
+        );
+
+        if (decision)
+        {
+            PlayerPrefs.DeleteAll();
+            SceneManager.LoadScene("MenuScene");
+        }
+                
     }
 
     private void ResetButtons()
@@ -160,4 +174,6 @@ public class MenuSelection : MonoBehaviour
             }
         }
     }
+
+
 }
